# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TrustPay Test Api",
        "version": "1.0.0"
    },
    "basePath": "/apitest",
    "paths": {
        "/customer/login": {
            "post": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "Customer Login",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "body",
                        "description": "Customer Details [CustomerObject]",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "Email Address",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "Password or PIN",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: customer_authenicated, data: [CustomerObject]}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
        "/seller/login": {
            "post": {
                "tags": [
                    "Seller Module"
                ],
                "summary": "Seller Login",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "body",
                        "description": "Seller Details [SellerObject]",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "Email Address",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "Password or PIN",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: seller_authenicated, data: [SellerObject]}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]}"
                    }
                }
            }
        },
...
        "/seller/changePassword": {
            "post": {
                "tags": [
                    "Seller Module"
                ],
                "summary": "Change Seller Password",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "body",
                        "description": "Change Seller password",
                        "required": true,
                        "schema": {
                            "properties": {
                                "current_password": {
                                    "description": "Seller Current Password",
                                    "type": "string"
                                },
                                "new_password": {
                                    "description": "Seller New Password",
                                    "type": "string"
                                },
                                "new_password_confirmation": {
                                    "description": "Seller New Password Confirmation",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: password_updated]}"
                    },
                    "201": {
                        "description": "{status: error, message: parameters missing or required/record_not_found/password_does_not_match}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]/role_not_defined}"
                    }
                }
            }
        },
        "/customer/changePassword": {
            "post": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "Change Customer Password",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "body",
                        "description": "Change customer password",
                        "required": true,
                        "schema": {
                            "properties": {
                                "current_password": {
                                    "description": "Customer Current Password",
                                    "type": "string"
                                },
                                "new_password": {
                                    "description": "Customer New Password",
                                    "type": "string"
                                },
                                "new_password_confirmation": {
                                    "description": "Customer New Password Confirmation",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status: success, message: password_updated]}"
                    },
                    "201": {
                        "description": "{status: error, message: parameters missing or required/record_not_found/password_does_not_match}"
                    },
                    "401": {
                        "description": "{status: error, message: invalid token/token expired, debug: [DebugInfoObject]/role_not_defined}"
                    }
                }
            }
        },
...
    },
...
}